+++ src/emmental/logging/__init__.py
@@ -1,14 +1,16 @@
 """Emmental logging module."""
 from emmental.logging.checkpointer import Checkpointer
+from emmental.logging.json_writer import JsonWriter
 from emmental.logging.log_writer import LogWriter
 from emmental.logging.logging_manager import LoggingManager
 from emmental.logging.tensorboard_writer import TensorBoardWriter
 from emmental.logging.wandb_writer import WandbWriter
 
 __all__ = [
+    "Checkpointer",
+    "JsonWriter",
     "LoggingManager",
     "LogWriter",
     "TensorBoardWriter",
-    "Checkpointer",
     "WandbWriter",
 ]
