+++ src/emmental/logging/wandb_writer.py
@@ -16,7 +16,15 @@ class WandbWriter(LogWriter):
         """Initialize TensorBoardWriter."""
         super().__init__()
 
-        # Set up wandb summary writer
+        # Set up wandb summary writer and save config
+        wandb.init(
+            project=Meta.config["logging_config"]["writer_config"][
+                "wandb_project_name"
+            ],
+            name=Meta.config["logging_config"]["writer_config"]["wandb_run_name"],
+            config=convert_to_serializable_json(copy.deepcopy(Meta.config)),
+        )
+
         self.write_config()
 
     def add_scalar_dict(
@@ -29,29 +37,3 @@ class WandbWriter(LogWriter):
           step: The current step.
         """
         wandb.log(metric_dict, step=step)
-
-    def write_config(self, config_filename: str = "config.yaml") -> None:
-        """Write the config to wandb and dump it to file.
-
-        Args:
-          config_filename: The config filename, defaults to "config.yaml".
-        """
-        wandb.init(
-            project=Meta.config["logging_config"]["writer_config"][
-                "wandb_project_name"
-            ],
-            name=Meta.config["logging_config"]["writer_config"]["wandb_run_name"],
-            config=convert_to_serializable_json(copy.deepcopy(Meta.config)),
-        )
-
-    def write_log(self, log_filename: str = "log.json") -> None:
-        """Dump the log to file.
-
-        Args:
-          log_filename: The log filename, defaults to "log.json".
-        """
-        pass
-
-    def close(self) -> None:
-        """Close the wandb writer."""
-        pass
